Moved functions out of GallerList to App.jsx so they can be passed as props
